+++ Modules/CalcSetup.lua
@@ -325,7 +325,7 @@ function calcs.initEnv(build, mode, override)
 			end
 		end
 	else
-		nodes = env.spec.allocNodes
+		nodes = copyTable(env.spec.allocNodes, true)
 	end
 	env.allocNodes = nodes
 
@@ -598,6 +598,16 @@ function calcs.initEnv(build, mode, override)
 		env.player.weaponData2 = env.player.itemList["Weapon 1"].weaponData[2]
 	else
 		env.player.weaponData2 = env.player.itemList["Weapon 2"] and env.player.itemList["Weapon 2"].weaponData and env.player.itemList["Weapon 2"].weaponData[2] or { }
+	end
+
+	-- Add granted passives
+	env.grantedPassives = { }
+	for _, passive in pairs(env.modDB:List(nil, "GrantedPassive")) do
+		local node = env.spec.tree.notableMap[passive]
+		if node then
+			nodes[node.id] = node
+			env.grantedPassives[node.id] = true
+		end
 	end
 
 	-- Merge modifiers for allocated passives