+++ Modules/Calcs.lua
@@ -398,7 +398,7 @@ function calcs.buildOutput(build, mode)
 			output.Minion.CurseList = output.CurseList
 		end
 
-		--infoDump(env)
+		infoDump(env)
 	end
 
 	return env